Registry code cleanup for Locust backend
